Conflicts:
	src/GPU/pair_lj_charmm_coul_long_gpu.cpp

+++ src/REAX/fix_reax_bonds.h
@@ -22,6 +22,7 @@ FixStyle(reax/bonds,FixReaxBonds)
 
 #include "stdio.h"
 #include "fix.h"
+#include "lmptype.h"
 
 namespace LAMMPS_NS {
 
@@ -39,7 +40,7 @@ class FixReaxBonds : public Fix {
   int nfreq;
   FILE *fp;
 
-  void OutputReaxBonds(int, FILE*);
+  void OutputReaxBonds(bigint, FILE*);
   int nint(const double&);
 };
